Split out handler files
